requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
langgraph
duckduckgo-search
python-dotenv
